chore: refactor and move out functions from vsphere into respective validators and fix field for rolePrivilegeRules in api
