+++ internal/vsphere/vsphere.go
@@ -3,23 +3,18 @@ package vsphere
 import (
 	"context"
 	"fmt"
-	"github.com/Knetic/govaluate"
 	"github.com/go-logr/logr"
 	"github.com/pkg/errors"
 	"github.com/spectrocloud-labs/valid8or-plugin-vsphere/api/v1alpha1"
 	"github.com/vmware/govmomi"
-	"github.com/vmware/govmomi/find"
 	"github.com/vmware/govmomi/object"
 	"github.com/vmware/govmomi/session"
 	"github.com/vmware/govmomi/session/keepalive"
 	"github.com/vmware/govmomi/vapi/rest"
-	"github.com/vmware/govmomi/vapi/tags"
 	"github.com/vmware/govmomi/vim25"
 	"github.com/vmware/govmomi/vim25/methods"
-	"github.com/vmware/govmomi/vim25/mo"
 	"github.com/vmware/govmomi/vim25/soap"
 	"net/url"
-	"sort"
 	"strings"
 	"sync"
 	"time"
@@ -31,10 +26,6 @@ var sessionCache = map[string]Session{}
 var sessionMU sync.Mutex
 var restClientLoggedOut = false
 
-type VMwareRolePrivilege struct {
-	rule       v1alpha1.RolePrivilegeValidationRule
-	Privileges map[string]bool
-}
 
 type VsphereCloudAccount struct {
 
@@ -62,13 +53,6 @@ type Session struct {
 type RulesEngine struct {
 	Driver *VSphereCloudDriver
 	Rules  []v1alpha1.RolePrivilegeValidationRule
-}
-
-type RegionZoneCategoryExistsInput struct {
-	Datacenter         string
-	Cluster            []string
-	RegionCategoryName string
-	ZoneCategoryName   string
 }
 
 type VSphereCloudDriver struct {
@@ -131,61 +115,6 @@ func (v *VSphereCloudDriver) GetVmwareUserPrivileges(userName string, authManage
 		}
 	}
 	return privileges, nil
-}
-
-func ToSlice(m map[string]bool) []interface{} {
-	values := make([]interface{}, 0, len(m))
-	for k, v := range m {
-		if v {
-			values = append(values, k)
-		}
-	}
-	return values
-}
-
-func (v *VMwareRolePrivilege) ValidateVMwareRolePrivilege() bool {
-	data := map[string]interface{}{
-		"vmware_user_privileges": ToSlice(v.Privileges),
-	}
-	for _, expr := range v.rule.Expressions {
-		expression, err := govaluate.NewEvaluableExpression(expr)
-		if err != nil {
-			return false
-		} else {
-			result, err := expression.Evaluate(data)
-			if err != nil {
-				return false
-			} else {
-				if result == false {
-					return false
-				}
-			}
-		}
-	}
-	return true
-}
-
-func IsValidRule(rule v1alpha1.RolePrivilegeValidationRule, privileges map[string]bool) bool {
-	// convert the keys of the map to a slice of strings
-	keys := make([]string, 0, len(privileges))
-	for k := range privileges {
-		keys = append(keys, k)
-	}
-
-	// sort the slice of keys
-	sort.Strings(keys)
-
-	if rule.IsEnabled {
-		switch rule.RuleType {
-		case "VMwareRolePrivilege":
-			rolePrivilegeRule := VMwareRolePrivilege{}
-			rolePrivilegeRule.rule = rule
-			rolePrivilegeRule.Privileges = privileges
-			return rolePrivilegeRule.ValidateVMwareRolePrivilege()
-		}
-	}
-
-	return false
 }
 
 func GetOrCreateSession(
@@ -311,94 +240,4 @@ func ClearCache(sessionKey string) {
 	sessionMU.Lock()
 	defer sessionMU.Unlock()
 	delete(sessionCache, sessionKey)
-}
-
-func RegionZoneCategoryExists(tagsManager *tags.Manager, finder *find.Finder, input RegionZoneCategoryExistsInput) (*bool, error) {
-	isTrue, isFalse := true, false
-	regionCategoryID, zoneCategoryID := "", ""
-
-	cats, err := tagsManager.GetCategories(context.TODO())
-	if err != nil {
-		return &isFalse, err
-	}
-	var regionZoneTags []tags.Category
-	for _, category := range cats {
-		switch category.Name {
-		case input.RegionCategoryName:
-			regionCategoryID = category.ID
-			regionZoneTags = append(regionZoneTags, category)
-		case input.ZoneCategoryName:
-			zoneCategoryID = category.ID
-			regionZoneTags = append(regionZoneTags, category)
-		}
-	}
-
-	if len(regionZoneTags) < 2 {
-		return &isFalse, nil
-	}
-
-	// check if datacenter has region tag
-	list, err := finder.ManagedObjectList(context.TODO(), fmt.Sprintf("/%s", input.Datacenter))
-	if err != nil {
-		return nil, err
-	}
-
-	// return early if no can't find the managedobject list
-	if len(list) == 0 {
-		return nil, nil
-	}
-	var refs []mo.Reference
-	refs = append(refs, list[0].Object.Reference())
-	attachedTags, err := tagsManager.GetAttachedTagsOnObjects(context.TODO(), refs)
-	if err != nil {
-		return nil, err
-	}
-	isDatacenterTaggedWithRegion := false
-
-	for _, attachedTag := range attachedTags {
-		for _, tagName := range attachedTag.Tags {
-			if tagName.CategoryID == regionCategoryID {
-				isDatacenterTaggedWithRegion = true
-				break
-			}
-		}
-	}
-
-	// check if all compute clusters has zone tag
-	areComputeClustersTaggedWithZone := true
-	for _, cluster := range input.Cluster {
-		list, err = finder.ManagedObjectList(context.TODO(), fmt.Sprintf("/%s/host/%s", input.Datacenter, cluster))
-		if err != nil {
-			return nil, err
-		}
-		// return early if no can't find the managedobject list
-		if len(list) == 0 {
-			return nil, nil
-		}
-		refs = nil
-		refs = append(refs, list[0].Object.Reference())
-		attachedTags, err := tagsManager.GetAttachedTagsOnObjects(context.TODO(), refs)
-		if err != nil {
-			return nil, err
-		}
-		found := false
-		for _, tag := range attachedTags {
-			if found {
-				break
-			}
-			for _, tagName := range tag.Tags {
-				if tagName.CategoryID == zoneCategoryID {
-					found = true
-					break
-				}
-			}
-		}
-		areComputeClustersTaggedWithZone = areComputeClustersTaggedWithZone && found
-	}
-
-	if areComputeClustersTaggedWithZone && isDatacenterTaggedWithRegion && len(regionZoneTags) >= 2 {
-		return &isTrue, nil
-	}
-
-	return &isFalse, nil
 }
